refactor : rename private field for code style and added comments

+++ include/interfaces/TapeInterface.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <iostream>
+
 class TapeInterface {
 public:
   virtual ~TapeInterface() = default;
@@ -10,5 +12,5 @@ public:
   virtual void moveRight() = 0;
   virtual void rewind() = 0;
   virtual bool isAtEnd() const = 0;
-  virtual int getSize() const = 0;
+  virtual size_t getSize() const = 0;
 };
